+++ CHANGELOG.md
@@ -5,9 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased] Unreleased
 
 ### Changed
+
 - Update code to be compatible with Codeception v5, PHP 8.0+ and Composer API 2.2+.
 - Move classes in the `tad\WPBrowser` namespace to the `lucatume\WPBrowser` one.
 - Move classes in the `Codeception` namespace to the `lucatume\WPBrowser` one.
+- Moved functions from the `lucatume\WPBrowser` namespace to dedicated classes in the `lucatume\WPBrowser\Utils`
+  namespace.
 
 ## [3.1.6] 2022-04-28;